+++ src/pages/ApplyEditPage.tsx
@@ -16,17 +16,17 @@ import DiscardModal from "../components/JD/DiscardModal";
 import JDContainer from "../components/JD/JDContainer";
 import ExperienceBox from "../components/JD/ExpContainer";
 import { ApplyAPI } from "../types/type";
-import { applypost } from "../services/jd";
+import { applyget, applypost } from "../services/jd";
 import { getCookie } from "../services/cookie";
 
-const JDWritePage: React.FC = () => {
+const ApplyEditPage: React.FC = () => {
   const [active, setActive] = useState(false); // 오른쪽 슬라이드 팝업 여부
   const [activebutton, setActivebutton] = useState(""); // 경험 분석 or 공고 보기
   const [applyData, setApplyData] = useState<ApplyAPI[]>([
     { question: "", answer: "" },
     { question: "", answer: "" },
   ]); //문항 데이터
-  const [editing, setEditing] = useState(true); //수정중 여부
+  const [editing, setEditing] = useState(false); //수정중 여부
   const [completed, setCompleted] = useState(false); //작성 완료
   const [isAllFilled, setIsAllFilled] = useState(false); // 문항이 빈칸이 없는지 검사
 
@@ -37,6 +37,7 @@ const JDWritePage: React.FC = () => {
   const nav = useNavigate();
   const jdId: string = useParams().jdId!; //공고 id
   const user = getCookie("user"); //토큰 받아오기용
+  const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
     window.scrollTo({
@@ -44,6 +45,9 @@ const JDWritePage: React.FC = () => {
       behavior: "auto",
     });
     // console.log("token", user.token);
+    if (jdId) {
+      getApplyData(jdId, user.token);
+    }
   }, []);
 
   //모든 질문이 다 채워졌는지 검사
@@ -128,6 +132,21 @@ const JDWritePage: React.FC = () => {
       setEditing(false);
       //api post 요청
     }
+  };
+
+  const getApplyData = async (jdId: string, token: string) => {
+    try {
+      const response = await applyget(jdId, token);
+      const mappedData = response.data.applyContentList.map((apply: any) => ({
+        question: apply.question,
+        answer: apply.answer,
+      }));
+      setApplyData(mappedData);
+    } catch (error) {
+      console.error(error);
+      alert(JSON.stringify(error));
+    }
+    setIsLoading(false);
   };
 
   //자기소개서 post api 요청
@@ -338,7 +357,7 @@ const JDWritePage: React.FC = () => {
   );
 };
 
-export default JDWritePage;
+export default ApplyEditPage;
 
 const StyledDivContainer = styled.div`
   width: 100%;
