+++ src/pages/JDDetailPage.tsx
@@ -31,7 +31,7 @@ const JDDetailPage: React.FC = () => {
     content: "",
     link: "",
     writeStatus: "",
-    createdAt: null,
+    createdAt: "",
     startAt: null,
     endedAt: null,
   });
@@ -51,9 +51,9 @@ const JDDetailPage: React.FC = () => {
 
   const handleNavigate = () => {
     if (firstTime) {
-      nav(`/jd/${jdId}`);
+      nav(`/jd/apply/${jdId}`);
     } else {
-      nav(`/jd/edit/${jdId}`);
+      nav(`/jd/apply/edit/${jdId}`);
     }
   };
 
@@ -73,9 +73,19 @@ const JDDetailPage: React.FC = () => {
     }
   }, [active]);
 
+  const formatDate = (createdAt: Date) => {
+    const date = new Date(createdAt);
+    const formattedDate = `${date.getDate()}.${date.getMonth() + 1}.${date
+      .getFullYear()
+      .toString()
+      .substring(2)}`;
+    return formattedDate;
+  };
+
   const getJobData = async (jdId: string, token: string) => {
     try {
       const response = await jobdescriptionget(jdId, token);
+      const FormatstartDate = formatDate(response.data.createdAt);
       const jdApiData: JobDescriptionAPI = {
         enterpriseName: response.data.enterpriseName,
         title: response.data.title,
@@ -83,7 +93,7 @@ const JDDetailPage: React.FC = () => {
         content: response.data.link,
         writeStatus: response.data.writeStatus,
         link: response.data.link,
-        createdAt: response.data.createdAt,
+        createdAt: FormatstartDate,
         startAt: response.data.startedAt,
         endedAt: response.data.endedAt,
       };
@@ -129,13 +139,25 @@ const JDDetailPage: React.FC = () => {
             </TopTitleBar>
             <JobContainer>
               <JobStatusBar>
-                {jdData.writeStatus !== "NOT_APPLIED" && (
-                  <StateBox
-                    className="job_status"
-                    status={jdData.writeStatus}
-                  />
-                )}
-                <div className="job_date">{jdData.createdAt?.toString()}</div>
+                <div className="left_container">
+                  {jdData.writeStatus !== "NOT_APPLIED" && (
+                    <StateBox
+                      className="job_status"
+                      status={jdData.writeStatus}
+                    />
+                  )}
+                  <div className="job_date">{jdData.createdAt?.toString()}</div>
+                </div>
+                <div className="right_container">
+                  <div
+                    onClick={() => {
+                      nav(`/jd/edit/${jdId}`);
+                    }}
+                  >
+                    수정
+                  </div>
+                  <div>삭제</div>
+                </div>
               </JobStatusBar>
               <JobTopBox>
                 <JobTopTitleBox>
@@ -340,16 +362,34 @@ const JobStatusBar = styled.div`
     display: flex;
     width: 100%;
     flex-direction: row;
+    justify-content: space-between;
     padding: 0.75rem 2rem;
     height: 3rem;
     background: ${(props) => props.theme.colors.neutral50};
     border-top-right-radius: 0.9rem;
     border-top-left-radius: 0.9rem;
     align-items: center;
+    .left_container{
+        display: flex;
+        flex-direction: row;
+        align-items: center;
     .job_date {
         color:  ${(props) => props.theme.colors.neutral500};
         ${(props) => props.theme.fonts.cap1};
         margin-left: 1rem;
+    }
+    }
+    .right_container{
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: center;
+        color:  ${(props) => props.theme.colors.neutral500};
+        font-size: 1rem;
+        font-style: normal;
+        gap: 1rem;
+        font-weight: 400;
+        text-decoration-line: underline;
     }
     
 `;
